+++ UnitTest/TestKxUf.hpp
@@ -14,7 +14,7 @@ class TestKxUf : public CxxTest::TestSuite
 
     void testConstructor0()
     {
-      const int size = 0;
+      const size_t size = 0;
       KxTYPE<Uf, size> stdfStr;
       Uf data[size];
       TS_ASSERT_SAME_DATA(stdfStr.mData, data, size);
@@ -28,7 +28,7 @@ class TestKxUf : public CxxTest::TestSuite
 
     void testConstructor1()
     {
-      const int size = 10;
+      const size_t size = 10;
       KxTYPE<Uf, size> stdfStr;
       Uf data[size];
       TS_ASSERT_SAME_DATA(stdfStr.mData, data, size);
@@ -42,7 +42,7 @@ class TestKxUf : public CxxTest::TestSuite
 
     void testConstructor2()
     {
-      const int size = 100;
+      const size_t size = 100;
       KxTYPE<Uf, size> stdfStr;
       Uf data[size];
       for(size_t i = 0; i < size; i++)
@@ -56,10 +56,10 @@ class TestKxUf : public CxxTest::TestSuite
       TS_ASSERT_EQUALS(stdfStr.size(Uf::TYPE_4B), 4*size);
       TS_ASSERT_EQUALS(stdfStr.size(Uf::TYPE_8B), 8*size);
       Uf data88 = stdfStr[88];
-      TS_ASSERT_EQUALS(data88.getValue(), 88);
+      TS_ASSERT_EQUALS(data88.getValue(), 88u);
       TS_ASSERT_EQUALS(data88.to_string(), data[88].to_string());
       Uf data99 = stdfStr[99];
-      TS_ASSERT_EQUALS(data99.getValue(), 99);
+      TS_ASSERT_EQUALS(data99.getValue(), 99u);
       TS_ASSERT_EQUALS(data99.to_string(), data[99].to_string());
       TS_ASSERT_EQUALS(stdfStr.to_string(), "0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99");
       stdfStr.clear();
@@ -70,7 +70,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead11.txt";
 
-      const int size = 1;
+      const size_t size = 1;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -93,7 +93,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead21.txt";
 
-      const int size = 1;
+      const size_t size = 1;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -116,7 +116,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead41.txt";
 
-      const int size = 1;
+      const size_t size = 1;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -139,7 +139,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead81.txt";
 
-      const int size = 1;
+      const size_t size = 1;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -162,7 +162,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead12.txt";
 
-      const int size = 10;
+      const size_t size = 10;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -185,7 +185,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead22.txt";
 
-      const int size = 10;
+      const size_t size = 10;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -208,7 +208,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead42.txt";
 
-      const int size = 10;
+      const size_t size = 10;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -231,7 +231,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead82.txt";
 
-      const int size = 10;
+      const size_t size = 10;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -254,7 +254,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead13.txt";
 
-      const int size = 20;
+      const size_t size = 20;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -283,7 +283,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead23.txt";
 
-      const int size = 20;
+      const size_t size = 20;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -312,7 +312,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead43.txt";
 
-      const int size = 20;
+      const size_t size = 20;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -341,7 +341,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteRead83.txt";
 
-      const int size = 20;
+      const size_t size = 20;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
@@ -370,7 +370,7 @@ class TestKxUf : public CxxTest::TestSuite
     {
       const char *filename = "TestKxUf.testWriteReadxx.txt";
 
-      const int size = 20;
+      const size_t size = 20;
       Uf data[size];
 
       KxTYPE<Uf, size> stdfStrIn;
