+++ UnitTest/TestVn.hpp
@@ -14,9 +14,9 @@ class TestVn : public CxxTest::TestSuite
 
     void testConstructor0()
     {
-      const int size = 0;
+      const size_t size = 0;
       Vn stdfStr;
-      TS_ASSERT_EQUALS(stdfStr.max_size(), 255);
+      TS_ASSERT_EQUALS(stdfStr.max_size(), 255u);
       TS_ASSERT_EQUALS(stdfStr.size(), size);
       TS_ASSERT_EQUALS(stdfStr.to_string(), "");
     }
@@ -37,8 +37,8 @@ class TestVn : public CxxTest::TestSuite
       Dn dn;
       JxN1<1> n1;
       stdfStr.addU1(u1).addU2(u2).addU4(u4).addI1(i1).addI2(i2).addI4(i4).addR4(r4).addR8(r8).addCn(cn).addBn(bn).addDn(dn).addN1(n1);
-      TS_ASSERT_EQUALS(stdfStr.max_size(), 255);
-      TS_ASSERT_EQUALS(stdfStr.size(), 49);
+      TS_ASSERT_EQUALS(stdfStr.max_size(), 255u);
+      TS_ASSERT_EQUALS(stdfStr.size(), 49u);
       TS_ASSERT_EQUALS(stdfStr.to_string(), "U1,0,B0,U2,0,B0,U4,0,I1,0,B0,I2,0,B0,I4,0,B0,R4,0,B0,R8,0,Cn,,Bn,,Dn,,N1,0");
     }
 
@@ -49,7 +49,7 @@ class TestVn : public CxxTest::TestSuite
       I2 i2 = 510;
       Cn cn = "AB";
       stdfStr.addCn(cn).addU1(u1).addI2(i2);
-      TS_ASSERT_EQUALS(stdfStr.size(), 10);
+      TS_ASSERT_EQUALS(stdfStr.size(), 10u);
       TS_ASSERT_EQUALS(stdfStr.to_string(), "Cn,AB,U1,255,B0,I2,510");
       stdfStr.clear();
       TS_ASSERT_EQUALS(stdfStr.to_string(), "");
@@ -77,7 +77,7 @@ class TestVn : public CxxTest::TestSuite
       ofstream outfile(filename, ofstream::binary);
       stdfStrIn.write(outfile);
       outfile.close();
-      TS_ASSERT_EQUALS(stdfStrIn.size(), 49)
+      TS_ASSERT_EQUALS(stdfStrIn.size(), 49u)
 
       Vn stdfStrOut;
       ifstream infile(filename, ifstream::binary);
@@ -110,7 +110,7 @@ class TestVn : public CxxTest::TestSuite
       ofstream outfile(filename, ofstream::binary);
       stdfStrIn.write(outfile);
       outfile.close();
-      TS_ASSERT_EQUALS(stdfStrIn.size(), 56)
+      TS_ASSERT_EQUALS(stdfStrIn.size(), 56u)
 
       Vn stdfStrOut;
       ifstream infile(filename, ifstream::binary);
